Documented NR test scripts.  Cleaned up some code.
