+++ tst/inputs/lwave_mhd.athinput
