+++ tst/test_suite/gr/test_gr_bondi_mpicpu.py
@@ -1,7 +1,8 @@
-# Automatic test based on GR Bondi Hydrodynamic Accretion
-#
-# Runs test in 3D with SMR and checks L1 error is below bound.  Errors are computed by
-# the executable automatically and stored in the temporary file gr_bondi-errs.dat.
+"""
+GR Bondi hydrodynamic accretion test.
+Runs test in 3D with SMR and checks L1 error is below bound.  Errors are computed by
+the executable automatically and stored in the temporary file gr_bondi-errs.dat.
+"""
 
 # Modules
 import sys
@@ -13,20 +14,19 @@ import scripts.utils.athena as athena
 import athena_read
 import numpy as np
 
-# Define run arguments
 def arguments():
+    """Assemble arguments for run command"""
     return [f'job/basename=gr_bondi',
              'time/tlim=50.0',
              'time/integrator=rk2',
              'hydro/reconstruct=plm']
 
-# Run AthenaK with MPI and arguments
 input_file = "inputs/gr_bondi.athinput"
 def test_run():
     """Run a single test with given arguments."""
     try:
         results = testutils.mpi_run(input_file, arguments(), threads=4,)
-        assert results, f"GR Bondi test failed./AthenaK did not run successfully."
+        assert results, f"GR Bondi test run failed."
         # Check the errors in the output
         data = athena_read.error_dat(f'gr_bondi-errs.dat')
         L1_RMS_INDEX = 4  # Index for L1 RMS error in data
