+++ tst/test_suite/gr/test_gr_lwave1d_cpu.py
@@ -1,11 +1,12 @@
-
-# Automatic test based on linear wave convergence in 1D
-# In hydro, both L-/R-going sound waves and the entropy wave are tested.
-# Note errors are very sensitive to the exact parameters (e.g. cfl_number, time limit)
-# used. For the hard-coded error limits to apply, run parameters must not be changed.
+"""
+Linear wave convergence test for general-relativistic hydro/MHD in 1D.
+Runs tests in both hydro and MHD for different
+  - time integrators
+  - reconstruction algorithms
+  - Riemann solvers
+"""
 
 # Modules
-
 import sys
 sys.path.append('../vis/python')
 sys.path.append('../tst/test_suite')
@@ -15,8 +16,8 @@ import scripts.utils.athena as athena
 import athena_read
 import numpy as np
 
-# Threshold errors and convergence rates
-# for different integrators, reconstructions, and wave types
+# Threshold errors and error ratios for different integrators, reconstruction,
+# algorithms, and wave types
 maxerrors={
     ('hydro', 'rk3', 'plm', '0'): (1.7e-08,0.28),
     ('hydro', 'rk3', 'ppm4', '0'): (6.8e-09,0.28),
@@ -59,8 +60,6 @@ maxerrors={
     ('mhd', 'rk3', 'ppmx', '3'): (1.4e-11,0.064),
     ('mhd', 'rk3', 'wenoz', '3'): (5.7e-12,0.033),}
 
-
-_soe = ['hydro', 'mhd']  # system of equations to test
 _int   = ['rk3']
 _recon = ['plm', 'ppm4', 'ppmx', 'wenoz']
 _wave={}
@@ -72,6 +71,7 @@ _flux['hydro'] = ['llf', 'hlle']
 _res = [32, 64] # resolutions to test
 
 def arguments(iv, rv, fv, wv, res, soe, name):
+    """Assemble arguments for run command"""
     return  [f'job/basename={name}',
             'time/tlim=1.0',
             'time/integrator=' + iv,
@@ -92,10 +92,11 @@ def arguments(iv, rv, fv, wv, res, soe, name):
 
 @pytest.mark.parametrize("iv" , _int)
 @pytest.mark.parametrize("rv" , _recon)
-@pytest.mark.parametrize("soe" , _soe)
+@pytest.mark.parametrize("soe" , ["hydro","mhd"])  # system of eqns to test
 def test_run(iv, rv, soe):
-    """Run tests of GR hydro and mhd."""
+    """Loop over Riemann solvers and run test with given integrator/resolution/physics."""
     for fv in _flux[soe]:
+        # returns error in L/R wave specified by 'left_wave'/'right_wave' arguments
         l1_rms_l,l1_rms_r = testutils.test_error_convergence(
             f"inputs/lwave_rel{soe}.athinput",
             f"gr_lwave_{soe}",
