+++ tst/test_suite/nr/test_nr_lwave2d_amr_cpu.py
@@ -1,11 +1,12 @@
-
-# Automatic test based on linear wave convergence in 2D
-# In hydro, both L-/R-going sound waves and the entropy wave are tested.
-# Note errors are very sensitive to the exact parameters (e.g. cfl_number, time limit)
-# used. For the hard-coded error limits to apply, run parameters must not be changed.
+"""
+Linear wave convergence test for non-relativistic hydro/MHD in 2D with AMR.
+Runs tests in both hydro and MHD for different
+  - time integrators
+  - reconstruction algorithms
+Only tests "0" wave and HLLC/HLLD Riemann solvers
+"""
 
 # Modules
-
 import sys
 sys.path.insert(0, '../vis/python')
 sys.path.insert(0, '../test_suite')
@@ -15,8 +16,8 @@ import scripts.utils.athena as athena
 import athena_read
 import numpy as np
 
-# Threshold errors and convergence rates
-# for different integrators, reconstructions, and wave types
+# Threshold errors and error ratios for different integrators, reconstruction,
+# algorithms, and waves
 errors = {
     ('hydro', 'rk2', 'plm', '0'): (6.9e-08,0.39),
     ('hydro', 'rk2', 'ppm4', '0'): (5.5e-08,0.37),
@@ -43,7 +44,7 @@ _flux = {'hydro': ['hllc'], 'mhd': ['hlld']}
 _res = [32, 64]                            # resolutions to test
 
 def arguments(iv,rv,fv,wv,res,soe,name):
-    """Run the Athena++ test with given parameters."""
+    """Assemble arguments for run command"""
     vflow = 1.0 if wv=='3' else 0.0
     return [f'job/basename={name}',
             'time/tlim=1.0',
@@ -66,10 +67,11 @@ def arguments(iv,rv,fv,wv,res,soe,name):
 @pytest.mark.parametrize("rv" , _recon)
 @pytest.mark.parametrize("soe",["hydro","mhd"])
 def test_run(iv, rv, soe):
-     for fv in _flux[soe]:
-        """Run a single test with given parameters."""
+    """Loop over Riemann solvers and run test with given integrator/resolution/physics."""
+    for fv in _flux[soe]:
+        # Ignore return arguments
         _,_ = testutils.test_error_convergence(
-            f"inputs/linear_wave_{soe}.athinput",
+            f"inputs/lwave_{soe}.athinput",
             f"lwave2d_amr_{soe}",
             arguments,
             errors,
