+++ tst/test_suite/nr/test_nr_rj2a_cpu.py
@@ -1,3 +1,11 @@
+"""
+RJ2a MHD shocktube test for non-relativistic MHD
+Runs tests for different
+  - reconstruction algorithms
+  - Riemann solvers
+"""
+
+# Modules
 import sys
 sys.path.insert(0, '../vis/python')
 sys.path.insert(0, '../test_suite')
@@ -9,11 +17,12 @@ import numpy as np
 
 
 _recon = ['plm','ppm4','ppmx','wenoz']  # do not change order
-_flux = ['llf', 'hlld']
+_flux = ['llf', 'hlle', 'hlld']
 _res  = [128,256]                            # resolutions to test
 input_file = "inputs/rj2a.athinput"
 
 def compute_error(data,tlim=0.2):
+    # compares density to analytic solution
     xfp = 2.2638*tlim
     xrp = (0.53432 + 1.0/np.sqrt(np.pi*1.309))*tlim
     xsp = (0.53432 + 0.48144/1.309)*tlim
@@ -31,8 +40,8 @@ def compute_error(data,tlim=0.2):
                   np.where(r > xfm, 1.4903, 1.08)))))))
     return (np.abs(data['dens'] - dens)).mean()
 
-
 def arguments(iv, rv, fv, res):
+    """Assemble arguments for run command"""
     return ['mesh/nx1='+repr(res),
             'meshblock/nx1=' + repr(np.min(_res)),
             'mesh/nghost=' + repr(2 if rv=='plm' else 3),
@@ -45,12 +54,12 @@ errors={}
 @pytest.mark.parametrize("rv" , _recon)
 @pytest.mark.parametrize("fv" , _flux)
 def test_run(fv, rv):
-    """Run a single test with given parameters."""
+    """Loop over resolutions and run test with given reconstruction/flux."""
     iv ="rk2" if rv=='plm' else "rk3"
     try:
         for res in _res:
             results = testutils.athenak_run(input_file,arguments(iv, rv, fv, res))
-            assert results, f"RJ2a test failed for iv={iv}, res={res}, fv={fv}, rv={rv}./AthenaK run did not complete successfully."
+            assert results, f"RJ2a test run failed for {iv}+{res}+{fv}+{rv}."
             data = athena_read.tab(f'tab/RJ2a.mhd_w.00001.tab')
             errors[res] = compute_error(data)
         # Check the errors in the output
@@ -60,8 +69,7 @@ def test_run(fv, rv):
         l1_rms_hr = errors[_res[1]]
         convrate = 0.6**(np.log2(_res[1] / _res[0]))
         if l1_rms_hr / l1_rms_lr > convrate:
-            pytest.fail(f"Not converging for  {iv}+{rv}+{fv} configuration, "
-                    f"conv: {l1_rms_hr / l1_rms_lr:g} threshold: {convrate:g}")
+            pytest.fail(f"Not converging for {iv}+{rv}+{fv} configuration, "
+                        f"conv: {l1_rms_hr / l1_rms_lr:g} threshold: {convrate:g}")
     finally:
-        print("Cleaning up test files...")
         testutils.cleanup()
