+++ tst/test_suite/nr/test_nr_sod_cpu.py
@@ -1,3 +1,11 @@
+"""
+Sod shocktube test for non-relativistic hydro
+Runs tests for different
+  - reconstruction algorithms
+  - Riemann solvers
+"""
+
+# Modules
 import sys
 sys.path.insert(0, '../vis/python')
 sys.path.insert(0, '../test_suite')
@@ -14,7 +22,7 @@ _res  = [128, 256]                            # resolutions to test
 input_file = "inputs/sod.athinput"
 
 def compute_error(data,tlim=0.25):
-    #Positions of shock, contact, head and foot of rarefaction for Sod test
+    # Positions of shock, contact, head and foot of rarefaction for Sod test
     xs = 1.7522*tlim
     xc = 0.92745*tlim
     xf = -0.07027*tlim
@@ -28,6 +36,7 @@ def compute_error(data,tlim=0.25):
 
 
 def arguments(iv, rv, fv, res):
+    """Assemble arguments for run command"""
     return [f'job/basename=sod',
             'mesh/nx1='+repr(res),
             'meshblock/nx1=' + repr(128),
@@ -41,12 +50,12 @@ errors={}
 @pytest.mark.parametrize("rv" , _recon)
 @pytest.mark.parametrize("fv" , _flux)
 def test_run(fv, rv):
-    """Run a single test with given parameters."""
+    """Loop over resolutions and run test with given reconstruction/flux."""
     iv ="rk2" if rv=='plm' else "rk3"
     try:
         for res in _res:
             results = testutils.athenak_run(input_file,arguments(iv, rv, fv, res))
-            assert results, f"Sod test failed for iv={iv}, res={res}, fv={fv}, rv={rv}./AthenaK run did not complete successfully."
+            assert results, f"Sod test run failed for {iv}+{res}+{fv}+{rv}."
             data = athena_read.tab(f'tab/sod.hydro_w.00001.tab')
             errors[res] = compute_error(data)
         # Check the errors in the output
@@ -57,8 +66,7 @@ def test_run(fv, rv):
         convrate = 0.6**(np.log2(_res[1] / _res[0]))
         if l1_rms_hr / l1_rms_lr > convrate:
             pytest.fail(f"Not converging for  {iv}+{rv}+{fv} configuration, "
-                    f"conv: {l1_rms_hr / l1_rms_lr:g} threshold: {convrate:g}")
+                        f"conv: {l1_rms_hr / l1_rms_lr:g} threshold: {convrate:g}")
     finally:
-        print("Cleaning up test files...")
         testutils.cleanup()
